lib: tenstorrent: remove fan_ctrl library
Replace internal fan_ctrl library with upstream PWM & sensor APIs
