+++ lib/tenstorrent/Kconfig
@@ -10,7 +10,6 @@ rsource "bh_chip/Kconfig"
 rsource "bist/Kconfig"
 rsource "boot_fs/Kconfig"
 rsource "event/Kconfig"
-rsource "fan_ctrl/Kconfig"
 rsource "fwupdate/Kconfig"
 rsource "jtag_bootrom/Kconfig"
 # zephyr-keep-sorted-stop
