--- lib/tenstorrent/fan_ctrl/CMakeLists.txt
@@ -1,4 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-zephyr_library()
-zephyr_library_sources(fan_ctrl.c)
